requirements.txt
django-widget-tweaks
unidecode